Move adapters file to a folder structure
